move graph components into a new file
