+++ src/widget.ts
@@ -8,8 +8,6 @@ import {
   DOMWidgetModel,
   DOMWidgetView,
   ISerializers,
-  WidgetModel,
-  WidgetView,
 } from '@jupyter-widgets/base';
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
@@ -20,7 +18,7 @@ import { MODULE_NAME, MODULE_VERSION } from './version';
 // Import the CSS
 import '../css/widget.css';
 
-import cytoscape, { EdgeSingular, NodeSingular } from 'cytoscape';
+import cytoscape from 'cytoscape';
 // @ts-ignore
 import cola from 'cytoscape-cola';
 // @ts-ignore
@@ -39,146 +37,6 @@ cytoscape.use(popper);
 cytoscape.use(dagre);
 cytoscape.use(klay);
 cytoscape.use(cola);
-
-export class ElementModel extends WidgetModel {
-  defaults() {
-    return {
-      ...super.defaults(),
-      _model_name: ElementModel.model_name,
-      _model_module: ElementModel.model_module,
-      _model_module_version: ElementModel.model_module_version,
-      _view_name: ElementModel.view_name,
-      _view_module: ElementModel.view_module,
-      _view_module_version: ElementModel.view_module_version,
-
-      group: '',
-      removed: false,
-      selected: false,
-      selectable: false,
-      locked: false,
-      grabbed: false,
-      grabbable: false,
-      classes: '',
-      data: {},
-      position: {},
-    };
-  }
-
-  static model_name = 'ElementModel';
-  static model_module = MODULE_NAME;
-  static model_module_version = MODULE_VERSION;
-  static view_name = 'ElementView';
-  static view_module = MODULE_NAME;
-  static view_module_version = MODULE_VERSION;
-}
-
-export class NodeModel extends ElementModel {
-  defaults() {
-    return {
-      ...super.defaults(),
-      _model_name: NodeModel.model_name,
-      _model_module: NodeModel.model_module,
-      _model_module_version: NodeModel.model_module_version,
-      _view_name: NodeModel.view_name,
-      _view_module: NodeModel.view_module,
-      _view_module_version: NodeModel.view_module_version,
-    };
-  }
-
-  static model_name = 'NodeModel';
-  static model_module = MODULE_NAME;
-  static model_module_version = MODULE_VERSION;
-  static view_name = 'NodeView';
-  static view_module = MODULE_NAME;
-  static view_module_version = MODULE_VERSION;
-}
-
-export class EdgeModel extends WidgetModel {
-  defaults() {
-    return {
-      ...super.defaults(),
-      _model_name: EdgeModel.model_name,
-      _model_module: EdgeModel.model_module,
-      _model_module_version: EdgeModel.model_module_version,
-      _view_name: EdgeModel.view_name,
-      _view_module: EdgeModel.view_module,
-      _view_module_version: EdgeModel.view_module_version,
-
-      group: '',
-      removed: false,
-      selected: false,
-      selectable: false,
-      locked: false,
-      grabbed: false,
-      grabbable: false,
-      classes: '',
-      data: {},
-      position: {},
-    };
-  }
-
-  static model_name = 'EdgeModel';
-  static model_module = MODULE_NAME;
-  static model_module_version = MODULE_VERSION;
-  static view_name = 'EdgeView';
-  static view_module = MODULE_NAME;
-  static view_module_version = MODULE_VERSION;
-}
-
-export class GraphModel extends WidgetModel {
-  defaults() {
-    return {
-      ...super.defaults(),
-      _model_name: 'GraphModel',
-      _model_module: GraphModel.model_module,
-      _model_module_version: GraphModel.model_module_version,
-      nodes: [],
-      edges: [],
-    };
-  }
-
-  static serializers: ISerializers = {
-    nodes: { deserialize: widgets.unpack_models },
-    edges: { deserialize: widgets.unpack_models },
-    ...WidgetModel.serializers,
-  };
-
-  static model_module = MODULE_NAME;
-  static model_module_version = MODULE_VERSION;
-
-  converts_dict() {
-    const graph: Array<any> = [];
-
-    for (let i = 0; i < this.attributes.nodes.length; i++) {
-      const node = this.attributes.nodes[i];
-      graph.push({
-        group: 'nodes',
-        data: node.get('data'),
-        selected: node.get('selected'),
-        selectable: node.get('selectable'),
-        locked: node.get('locked'),
-        grabbed: node.get('grabbed'),
-        classes: node.get('classes'),
-        position: node.get('position'),
-      });
-    }
-    for (let j = 0; j < this.attributes.edges.length; j++) {
-      const edge = this.attributes.edges[j];
-      graph.push({
-        group: 'edges',
-        data: edge.get('data'),
-        selected: edge.get('selected'),
-        selectable: edge.get('selectable'),
-        locked: edge.get('locked'),
-        grabbed: edge.get('grabbed'),
-        classes: edge.get('classes'),
-        position: edge.get('position'),
-      });
-    }
-
-    return graph;
-  }
-}
 
 export class CytoscapeModel extends DOMWidgetModel {
   defaults() {
@@ -217,86 +75,6 @@ export class CytoscapeModel extends DOMWidgetModel {
   static view_module_version = MODULE_VERSION;
 }
 
-export class ElementView extends WidgetView {
-  cytoscapeView: CytoscapeView;
-  private elem: NodeSingular | EdgeSingular;
-
-  constructor(params: any) {
-    super({
-      model: params.model,
-      options: params.options,
-    });
-    this.cytoscapeView = this.options.cytoscapeView;
-    const cyId = this.model.get('data')['id'];
-    this.elem = this.cytoscapeView.cytoscape_obj.getElementById(cyId);
-
-    this.model.on('change:group', this.valueChanged, this);
-    this.model.on('change:removed', this.valueChanged, this);
-    this.model.on('change:selected', this.valueChanged, this);
-    this.model.on('change:locked', this.valueChanged, this);
-    this.model.on('change:grabbed', this.valueChanged, this);
-    this.model.on('change:grabbable', () => {
-      const oy = this.elem as any;
-      oy.grabbable(this.model.get('grabbable'));
-    });
-    this.model.on('change:classes', this._updateClasses, this);
-    this.model.on('change:data', this._updateData, this);
-    this.model.on('change:position', this.valueChanged, this);
-  }
-
-  private _updateData() {
-    this.elem.data(this.model.get('data'));
-  }
-
-  private _updateClasses() {
-    this.elem.classes(this.model.get('classes'));
-  }
-
-  valueChanged() {
-    this.cytoscapeView.value_changed();
-  }
-}
-export class NodeView extends ElementView {
-  constructor(params: any) {
-    super({
-      model: params.model,
-      options: params.options,
-    });
-  }
-}
-
-export class EdgeView extends WidgetView {
-  cytoscapeView: CytoscapeView;
-  private elem: EdgeSingular;
-
-  constructor(params: any) {
-    super({
-      model: params.model,
-      options: params.options,
-    });
-    this.cytoscapeView = this.options.cytoscapeView;
-
-    this.model.on('change:group', this.valueChanged, this);
-    this.model.on('change:removed', this.valueChanged, this);
-    this.model.on('change:selected', this.valueChanged, this);
-    this.model.on('change:locked', this.valueChanged, this);
-    this.model.on('change:grabbed', this.valueChanged, this);
-    this.model.on('change:grabbable', this.valueChanged, this);
-    this.model.on('change:classes', this._updateClasses, this);
-    this.model.on('change:data', this.valueChanged, this);
-    this.model.on('change:position', this.valueChanged, this);
-    const cyId = this.model.get('data')['id'];
-    this.elem = this.cytoscapeView.cytoscape_obj.getElementById(cyId);
-  }
-  private _updateClasses() {
-    this.elem.classes(this.model.get('classes'));
-  }
-
-  valueChanged() {
-    this.cytoscapeView.value_changed();
-  }
-}
-
 export class CytoscapeView extends DOMWidgetView {
   cytoscape_obj: any;
   is_rendered = false;
@@ -307,21 +85,22 @@ export class CytoscapeView extends DOMWidgetView {
   render() {
     this.el.classList.add('custom-widget');
 
-    this.nodeViews = new widgets.ViewList(
-      this.addNodeModel,
-      this.removeNodeView,
-      this
-    );
-    this.nodeViews.update(this.model.get('graph').get('nodes'));
+    this.displayed.then(() => {
+      this.init_render();
+      this.nodeViews = new widgets.ViewList(
+        this.addNodeModel,
+        this.removeNodeView,
+        this
+      );
+      this.nodeViews.update(this.model.get('graph').get('nodes'));
 
-    this.edgeViews = new widgets.ViewList(
-      this.addEdgeModel,
-      this.removeEdgeView,
-      this
-    );
-    this.edgeViews.update(this.model.get('graph').get('edges'));
-
-    this.value_changed();
+      this.edgeViews = new widgets.ViewList(
+        this.addEdgeModel,
+        this.removeEdgeView,
+        this
+      );
+      this.edgeViews.update(this.model.get('graph').get('edges'));
+    });
 
     this.model.get('graph').on('change:nodes', this.value_changed, this);
     this.model.get('graph').on('change:edges', this.value_changed, this);
@@ -366,10 +145,6 @@ export class CytoscapeView extends DOMWidgetView {
     if (layout !== null) {
       layout.on_some_change(['width', 'height'], this._resize, this);
     }
-
-    this.displayed.then(() => {
-      this.init_render();
-    });
   }
 
   value_changed() {
@@ -415,7 +190,9 @@ export class CytoscapeView extends DOMWidgetView {
   }
 
   init_render() {
+    console.log('in init render');
     if (this.model.get('graph') !== null) {
+      console.log('in if in init_render');
       this.is_rendered = true;
       this.cytoscape_obj = cytoscape({
         container: this.el,
